GPU: store VRAM as [u16] instead of [u8] because it makes everything simpler except for 24bpp display, which is not too bad to handle

+++ ps1-core/src/api.rs
@@ -62,7 +62,7 @@ pub trait Renderer {
     /// # Errors
     ///
     /// Should propagate any error encountered while rendering the frame.
-    fn render_frame(&mut self, vram: &[u8], params: RenderParams) -> Result<(), Self::Err>;
+    fn render_frame(&mut self, vram: &[u16], params: RenderParams) -> Result<(), Self::Err>;
 }
 
 pub trait AudioOutput {
